DefaultCountingOutRhymer() class refactored
