+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,39 +1,37 @@
 package edu.kis.vh.nursery;
 
 public class DefaultCountingOutRhymer {
-    private static final int minusOne = -1;
-    private static final int eleven = 11;
-    private static final int twelve = 12;
-    private final int[] numbers = new int[twelve];
-    private int total = minusOne;
+    IntArrayStack intArrayStack = null;
+
+    public DefaultCountingOutRhymer() {
+        intArrayStack = new IntArrayStack();
+    }
+
+    public DefaultCountingOutRhymer(IntArrayStack intArrayStack) {
+        this.intArrayStack = intArrayStack;
+    }
 
     public int getTotal() {
-        return total;
+        return intArrayStack.getTotal();
     }
 
     public void countIn(int in) {
-        if (!isFull())
-            numbers[++total] = in;
+        intArrayStack.countIn(in);
     }
 
     public boolean callCheck() {
-        return total == minusOne;
+        return intArrayStack.callCheck();
     }
 
     public boolean isFull() {
-        return total == eleven;
+        return intArrayStack.isFull();
     }
 
-    protected int peekaboo() {
-        if (callCheck())
-            return minusOne;
-        return numbers[total];
+    public int peekaboo() {
+        return intArrayStack.peekaboo();
     }
 
     public int countOut() {
-        if (callCheck())
-            return minusOne;
-        return numbers[total--];
+        return intArrayStack.countOut();
     }
-
 }
